Hybrid of original groupby approach and new approach
I think this is good because it takes advantage of the performance
benefits of using `groupby` to do the calculations and the readability
benefits of creating a dataframe from scratch to map the old column
names to the new column names. I think a good next step in readability
would be to create an actual map (dict) of old column names to new
column names!